hypertension-nlp/pyproject.toml
[project]
name = "tedla-hypertension"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "invoke>=2.2.1",
    "ipykernel>=7.2.0",
    "ipython>=9.10.0",
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
    "markdown>=3.10.2",
    "medspacy-quickumls>=3.2",
    "nbconvert>=7.17.0",
    "numpy==1.26.4",
    "pandas>=2.3.3",
    "pip>=26.0.1",
    "polars>=1.38.1",
    "pyarrow>=23.0.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "spacy>=3.8.11",
    "tabulate>=0.9.0",
    "weasyprint>=68.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"


